+++ pyvo/utils/http.py
@@ -6,7 +6,7 @@ import platform
 import requests
 from ..version import version
 
-DEFAULT_USER_AGENT = 'python-pyvo/{} ({})'.format(version, platform.system())
+DEFAULT_USER_AGENT = f'pyVO/{version} Python/{platform.python_version()} ({platform.system()})'
 
 
 def use_session(session):